Added ability to create and use your own HelpGenerator
