+++ CLAP/HelpInfo.cs
@@ -3,44 +3,44 @@ using System.Collections.Generic;
 
 namespace CLAP
 {
-    internal class HelpInfo
+    public class HelpInfo
     {
-        internal List<ParserHelpInfo> Parsers { get; set; }
+        public List<ParserHelpInfo> Parsers { get; set; }
     }
 
-    internal class ParserHelpInfo
+    public class ParserHelpInfo
     {
-        internal Type Type { get; set; }
-        internal List<VerbHelpInfo> Verbs { get; set; }
-        internal List<GlobalParameterHelpInfo> Globals { get; set; }
+        public Type Type { get; set; }
+        public List<VerbHelpInfo> Verbs { get; set; }
+        public List<GlobalParameterHelpInfo> Globals { get; set; }
     }
 
-    internal class VerbHelpInfo
+    public class VerbHelpInfo
     {
-        internal bool IsDefault { get; set; }
-        internal string Description { get; set; }
-        internal List<string> Names { get; set; }
-        internal List<ParameterHelpInfo> Parameters { get; set; }
-        internal List<string> Validations { get; set; }
+        public bool IsDefault { get; set; }
+        public string Description { get; set; }
+        public List<string> Names { get; set; }
+        public List<ParameterHelpInfo> Parameters { get; set; }
+        public List<string> Validations { get; set; }
     }
 
-    internal class ParameterHelpInfo
+    public class ParameterHelpInfo
     {
-        internal bool Required { get; set; }
-        internal List<string> Names { get; set; }
-        internal List<string> Validations { get; set; }
-        internal Type Type { get; set; }
-        internal object Default { get; set; }
-        internal string Description { get; set; }
-        internal string Separator { get; set; }
+        public bool Required { get; set; }
+        public List<string> Names { get; set; }
+        public List<string> Validations { get; set; }
+        public Type Type { get; set; }
+        public object Default { get; set; }
+        public string Description { get; set; }
+        public string Separator { get; set; }
     }
 
-    internal class GlobalParameterHelpInfo
+    public class GlobalParameterHelpInfo
     {
-        internal List<string> Names { get; set; }
-        internal List<string> Validations { get; set; }
-        internal Type Type { get; set; }
-        internal string Description { get; set; }
-        internal string Separator { get; set; }
+        public List<string> Names { get; set; }
+        public List<string> Validations { get; set; }
+        public Type Type { get; set; }
+        public string Description { get; set; }
+        public string Separator { get; set; }
     }
 }
